rabbitmq_timeout/requirements.txt
pika==1.0.1
attrs==19.1.0
kombu==4.5.0
timeout-decorator==0.4.1
amqp==2.4.2

psycopg2==2.8.2
SQLAlchemy==1.3.3
#sanic==19.3.1
